// src/app/shared/components/member/enums/member-search-filter.enum.ts
export enum DefaultMemberSearchFilter {
  PageSize = 50,
  PageNumber = 0,
  RequestInt = -1,
}

export enum MemberSortDescriptors {
  LastNameAsc = 'lastName ASC',
  LastNameDesc = 'lastName DESC',
  FirstNameAsc = 'firstName ASC',
  FirstNameDesc = 'firstName DESC',
}

export enum FirstNameFilterIds {
  Id = '43033c03-8172-46fd-b695-227ea51d5f00',
  SelectedFilterFieldId = 'E3ECA7FF-C785-4329-B6D9-D4D554E22D0E',
}

export enum FirstNameFilterFields {
  FilterOrder = 0,
  SelectedAndOr = 0,
  SelectedAcceptedValue = 0,
  SelectedComparisonOperator = 7,
}

export enum LastNameFilterIds {
  Id = '43033c03-8172-46fd-b695-227ea51d5f00',
  SelectedFilterFieldId = '39b30b4e-af48-4727-b1b5-9e02e8baa0d3',
}

export enum LastNameFilterFields {
  FilterOrder = 1,
  SelectedAndOr = 2,
  SelectedAcceptedValue = 0,
  SelectedComparisonOperator = 7,
}

export enum PhoneNumberFilterIds {
  Id = '43033c03-8172-46fd-b695-227ea51d5f00',
  SelectedFilterFieldId = '0F87B9DC-1E97-4487-A969-3FFB2ADE2DF2',
}

export enum PhoneNumberFilterFields {
  FilterOrder = 2,
  SelectedAndOr = 2,
  SelectedAcceptedValue = 0,
  SelectedComparisonOperator = 7,
}